+++ src/Objects/Particles/G_PS_Star.h
@@ -16,7 +16,7 @@ public:
     - rotate_axis - axises for rotation of particles.
     - _color_start, _color_end - range for colors (each particle will have random value from that range).
     */
-    G_PS_Star(unsigned long particles_count, const ZC_Vec3<float>& _star_pos, float _radius, float _speed = 1.f, const ZC_Vec3<float>& rotate_axises = { 1.f, 0.f, 1.f },
+    G_PS_Star(unsigned long long particles_count, const ZC_Vec3<float>& _star_pos, float _radius, float _speed = 1.f, const ZC_Vec3<float>& rotate_axises = { 1.f, 0.f, 1.f },
         const ZC_Vec3<unsigned char>& _color_start = { 255, 100, 0 }, const ZC_Vec3<unsigned char>& _color_end = { 255, 255, 0 });
 
         // - _start_sphere_radius - radius in wich start to apear particles_sphere (only half of start sphere are in _length).
