+++ src/Objects/Particles/G_PS_Wind.cpp
@@ -4,7 +4,7 @@
 #include <ZC/Tools/Math/ZC_Vec.h>
 #include <ZC/Tools/ZC_Random.h>
 
-G_PS_Wind::G_PS_Wind(unsigned long particles_count, const ZC_Vec3<float>& _start_pos, float start_width, float start_height,
+G_PS_Wind::G_PS_Wind(unsigned long long particles_count, const ZC_Vec3<float>& _start_pos, float start_width, float start_height,
         const ZC_Vec3<float>& _wind_dir, float _length, float _speed, const ZC_Vec3<unsigned char>& color_start, const ZC_Vec3<unsigned char>& color_end)
     : G_ParticlesDrawer(particles_count, G_DL_AlphaBlending_ParticlesWind, 3),
     wind_pos(_start_pos),
