+++ src/Objects/Particles/G_PS_Wind.h
@@ -18,7 +18,7 @@ public:
     - speed - speed coef, default (1 meter(float) / 1 second).
     - _color_start, _color_end - range for colors (each particle will have random value from that range).
     */
-    G_PS_Wind(unsigned long particles_count, const ZC_Vec3<float>& _start_pos, float start_width, float start_height, const ZC_Vec3<float>& _wind_dir, float _length,
+    G_PS_Wind(unsigned long long particles_count, const ZC_Vec3<float>& _start_pos, float start_width, float start_height, const ZC_Vec3<float>& _wind_dir, float _length,
         float _speed = 1.f, const ZC_Vec3<unsigned char>& _color_start = { 175, 238, 238 }, const ZC_Vec3<unsigned char>& _color_end = { 0, 206, 209 });
 
         // - _start_sphere_radius - radius in wich start to apear particles (only half of start sphere are in _length).
